SAMSON-431 samsonPop and samsonPush addapted to the new network 2.1 scenario. Fix minor bugs in network 2.1
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@2506 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/samson/stream/PopQueueTask.h
@@ -58,12 +58,12 @@ namespace samson {
             friend class QueueTaskManager;
             friend class StreamManager;
             
-            int fromId ;            // Identifier of delilah to send packets back
+            size_t fromId ;         // Identifier of delilah to send packets back
             std::string queue;      // Name of the queue we are streaming out
             
         public:
             
-            StreamOutQueueTask( size_t _id , int _fromId, std::string _queue );
+            StreamOutQueueTask( size_t _id , size_t _fromId, std::string _queue );
             void sendMessage( engine::Buffer *buffer );
             void run();
